feat(iterables): allow overriding  builtins implementation
Iteralbes are compiled to JS generators with little room for easy experimention of different implimentations. It is now possible to to have loxia compile these constructs to a simple step function which is passed to an imported builtin function (_map_, _filter_, _while_, _until_, _fold_, _unfold_), along with some metadata about accu usage, result spreading and asyncronocity.

In the future this mechanism will be used for other functionality that can be expressed as runtime functions rather than compiling it to inline code.